% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{create_df_eod_raw}
\alias{create_df_eod_raw}
\title{Create a data frame containing the raw end-of-day data.}
\usage{
create_df_eod_raw(database)
}
\arguments{
\item{database}{Directory path and file name of the database}
}
\value{
A data frame containing the raw end-of-day data
}
\description{
Create a data frame containing the raw end-of-day data.
}
