% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{create_df_mid_raw}
\alias{create_df_mid_raw}
\title{Create a data frame containing the raw market index data.}
\usage{
create_df_mid_raw(database)
}
\arguments{
\item{database}{Directory path and file name of the database}
}
\value{
A data frame containing the raw market index data
}
\description{
Create a data frame containing the raw market index data.
}
