% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{distr_year.df}
\alias{distr_year.df}
\title{Distribution of the volume of data on year.}
\usage{
distr_year.df(tad_raw.df)
}
\arguments{
\item{tad_raw.df}{A data frame containing transaction data, i.e. end-of-day
data or market index data}
}
\value{
A data frame containing the year and the number of data entries
        in that year.
}
\description{
Distribution of the volume of data on year.
}
