% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{distr_year_symbol.df}
\alias{distr_year_symbol.df}
\title{Distribution of the volume of data on year and symbol.}
\usage{
distr_year_symbol.df(tad_raw.df)
}
\arguments{
\item{tad_raw.df}{A dataframe containing transaction data, i.e. end-of-day
data or market index data}
}
\description{
Distribution of the volume of data on year and symbol.
}
