% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{distr_year_symbols.barplot}
\alias{distr_year_symbols.barplot}
\title{Barplot of the number of symbols per year.}
\usage{
distr_year_symbols.barplot(tad_raw.df)
}
\arguments{
\item{tad_raw.df}{A data frame containing transaction data, i.e. end-of-day
data or market index data}
}
\description{
Barplot of the number of symbols per year.
}
