% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{distr_year_trading_days.barplot}
\alias{distr_year_trading_days.barplot}
\title{Barplot of the number of trading days per year.}
\usage{
distr_year_trading_days.barplot(tad_raw.df)
}
\arguments{
\item{tad_raw.df}{A data frame containing transaction data, i.e. end-of-day
data or market index data}
}
\description{
Barplot of the number of trading days per year.
}
