% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{eod_raw.df}
\alias{eod_raw.df}
\title{Create a data frame containing the raw end-of-day data.}
\usage{
eod_raw.df(database)
}
\arguments{
\item{database}{Directory path and file name of the database}
}
\value{
A data frame containing the raw end-of-day data
}
\description{
Create a data frame containing the raw end-of-day data.
}
