% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{get_distr_year}
\alias{get_distr_year}
\title{Distribution of the volume of data on year.}
\usage{
get_distr_year(df_tad)
}
\arguments{
\item{df_tad}{A dataframe containing transaction data}
}
\description{
Distribution of the volume of data on year.
}
