% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppr_utils.R
\name{get_distr_year_symbol}
\alias{get_distr_year_symbol}
\title{Distribution of the volume of data on year and symbol.}
\usage{
get_distr_year_symbol(df_tad)
}
\arguments{
\item{df_tad}{A dataframe containing transaction data}
}
\description{
Distribution of the volume of data on year and symbol.
}
