% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{get_packages}
\alias{get_packages}
\title{Create a combined and sorted character vector containing the necessary
package names.}
\usage{
get_packages(add_on_1 = vector(), add_on_2 = vector())
}
\arguments{
\item{add_on_1}{Character vector containing additional pacckage names.}

\item{add_on_2}{Character vector containing additional pacckage names.}
}
\description{
Create a combined and sorted character vector containing the necessary
package names.
}
