% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{install}
\alias{install}
\title{Install all packages relevant for sppr and show the resulting installation
state.}
\usage{
install(library = .libPaths()[1], repository = cran_repository)
}
\arguments{
\item{library}{character vector giving the library directories where to
install the packages. If missing, defaults to the first
element of .libPaths()}

\item{repository}{character vector, the base URL(s) of the repositories to
use, e.g., the URL of a CRAN mirror such as
"https://cloud.r-project.org". If missing, defaults to
"https://cran.r-project.org/"}
}
\description{
Install all packages relevant for sppr and show the resulting installation
state.
}
