% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_required}
\alias{install_required}
\title{Install the required packages and show the current installation state.}
\usage{
install_required()
}
\description{
Install the required packages and show the current installation state.
}
