% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{mid_range_symbol.df}
\alias{mid_range_symbol.df}
\title{Create a data frame containing the market index trading day range per symbol.}
\usage{
mid_range_symbol.df(database)
}
\arguments{
\item{database}{Directory path and file name of the database}
}
\value{
A data frame containing the raw end-of-day data
}
\description{
Create a data frame containing the market index trading day range per symbol.
}
