% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{remove}
\alias{remove}
\title{Remove packages and show the resulting installation state.}
\usage{
remove(library = .libPaths()[1])
}
\arguments{
\item{library}{character vector giving the library directories where to
remove the packages. If missing, defaults to the first
element of .libPaths()}
}
\description{
Remove packages and show the resulting installation state.
}
