% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_required}
\alias{remove_required}
\title{Remove required packages and show the current installation state.}
\usage{
remove_required()
}
\description{
Remove required packages and show the current installation state.
}
