% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{required_packages}
\alias{required_packages}
\title{Definition of the required packages.}
\usage{
required_packages(add_ons = vector())
}
\arguments{
\item{add_ons}{Character vvector containing additional pacckage names.}
}
\description{
Definition of the required packages.
}
