% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_installed}
\alias{show_installed}
\title{Show the status of the required packages.}
\usage{
show_installed(library = "packrat/lib")
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Show the status of the required packages.
}
