% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_specific}
\alias{show_specific}
\title{Show the status of all packages relevant for sppr.}
\usage{
show_specific(library = .libPaths()[1])
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Show the status of all packages relevant for sppr.
}
