% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppr_install.R
\name{sppr_install}
\alias{sppr_install}
\title{Install the required packages from the Swiss CRAN network.}
\usage{
sppr_install()
}
\description{
Install the required packages from the Swiss CRAN network.
}
