% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install}
\alias{install}
\title{Install all packages relevant for sppr and show the resulting installation
state.}
\usage{
install(library = .libPaths()[1])
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Install all packages relevant for sppr and show the resulting installation
state.
}
