% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{remove}
\alias{remove}
\title{Remove packages and show the resulting installation state.}
\usage{
remove(library = .libPaths()[1])
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Remove packages and show the resulting installation state.
}
