% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_required}
\alias{remove_required}
\title{Remove required packages and show the current installation state.}
\usage{
remove_required(library = .libPaths()[1])
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Remove required packages and show the current installation state.
}
