% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{show_packages}
\alias{show_packages}
\title{Show the status of all packages relevant for sppr.}
\usage{
show_packages(library = .libPaths()[1])
}
\arguments{
\item{library}{character vector giving the library directories where to
search for the packages. If missing, defaults to the first
element of .libPaths()}
}
\description{
Show the status of all packages relevant for sppr.
}
