% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{update}
\alias{update}
\title{Update all packages and show the resulting installation state.}
\usage{
update(library = .libPaths()[1])
}
\arguments{
\item{library}{Character vector describing the location of R library
trees to search through}
}
\description{
Update all packages and show the resulting installation state.
}
